+++ trixnity-clientserverapi/trixnity-clientserverapi-client/src/commonMain/kotlin/de/connect2x/trixnity/clientserverapi/client/DownloadLimitExceededException.kt
@@ -7,9 +7,10 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 class DownloadLimitExceededException(
     val maxSize: Long,
     message: String? = null,
-    cause: Throwable? = null) : IllegalStateException("File could not be downloaded because it would exceed the limit of $maxSize bytes"),
+    cause: Throwable? = null
+) : IllegalStateException("File could not be downloaded because it would exceed the limit of $maxSize bytes"),
     CopyableThrowable<DownloadLimitExceededException> {
     override fun createCopy(): DownloadLimitExceededException {
-        return DownloadLimitExceededException(maxSize, message, this)
+        return DownloadLimitExceededException(maxSize, message, cause)
     }
 }