+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SignedPersistentDataUnit<T>` type alias
 - `SignedPersistentStateDataUnit<T>` type alias
 - `SignedPersistentMessageDataUnit<T>` type alias
+- `MediaService`: `getMedia`, `getEncryptedMedia` and `getThumbnail` without required `maxSize` parameter
 
 ### Removed